chore(utils): split into files
